Change include paths for moved headers.
