+++ opm/autodiff/SimulatorIncompTwophaseAd.cpp
@@ -35,7 +35,7 @@
 #include <opm/core/pressure/flow_bc.h>
 
 #include <opm/core/simulator/SimulatorReport.hpp>
-#include <opm/core/simulator/SimulatorTimer.hpp>
+#include <opm/simulators/timestepping/SimulatorTimer.hpp>
 #include <opm/core/utility/StopWatch.hpp>
 #include <opm/core/utility/DataMap.hpp>
 #include <opm/simulators/vtk/writeVtkData.hpp>
