+++ opm/simulators/timestepping/SimulatorTimer.cpp
@@ -18,7 +18,7 @@
 */
 
 #include "config.h"
-#include <opm/core/simulator/SimulatorTimer.hpp>
+#include <opm/simulators/timestepping/SimulatorTimer.hpp>
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
 #include <opm/parser/eclipse/Units/Units.hpp>
 #include <ostream>
